requirements.txt
langchain
line-bot-sdk
flask
numexpr
langchain_google_genai
